Add files via upload
Markdown Editor
Make a handy and useful tool that uses the markdown language!
Markdown is a special plain text formatting language that is extremely popular among developers. It is used in documents, research articles, Github README files, and other things. In this project, you will write an editor that will be able to recognize several tags, structures, and save your results to a file.

Completed Markdown Editor at https://hyperskill.org
The code that passes the Project Completion