Knight's Tour Puzzle. Stage 6/6: Give me an answer
Knight's Tour Puzzle
The Knight’s Tour is a fun puzzle where you move the knight so that it visits every square of the chessboard once. The rules are simple and fun, but the game is really hard to master. Let’s make a program to find a solution.
Learning outcomes: You will write a program designed to find a solution to the problem with a board of any size. You will learn about matrices, backtracking, recursion, and algorithms.
Completed Knight's Tour Puzzle at https://hyperskill.org
The code that passes the Project Completion